raspberry/raspberry_requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
fonttools==4.57.0
h11==0.14.0
idna==3.10
imageio==2.37.0
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
tifffile==2025.3.30
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
